adapt windows tracert
